Refactor QueryExpression to TS
