--- src/query/expression.js
@@ -1,22 +1,0 @@
-import { isObject } from '../lib/objects';
-
-export default class QueryExpression {
-  constructor(op, args) {
-    this.__oqe__ = true;
-    this.op = op;
-    this.args = args;
-  }
-
-  toString() {
-    const formattedArgs = this.args.map(arg => '' + arg).join(', ');
-    return `${this.op}(${formattedArgs})`;
-  }
-}
-
-export function queryExpression(op, ...args) {
-  return new QueryExpression(op, args);
-}
-
-export function isQueryExpression(obj) {
-  return isObject(obj) && obj.__oqe__;
-}
