+++ test/query/builder-test.js
@@ -1,4 +1,4 @@
-import { queryExpression as oqe } from '../../src/query/expression';
+import { queryExpression as oqe } from '../../src/query-expression';
 import qb from '../../src/query/builder';
 import '../test-helper';
